fix: App crashes deleting firebase image if local
If the image is local it can't delete the firebase version since it
doesn't exist.

+++ app/src/main/java/com/michaelvescovo/android/itemreaper/items/ItemsFragment.java
@@ -212,9 +212,11 @@ public class ItemsFragment extends Fragment implements ItemsContract.View {
 
     private void removeImageFromFirebase() {
         if (mDeletedItem.getImageUrl() != null) {
-            StorageReference photoRef = mFirebaseStorage.getReferenceFromUrl(
-                    mDeletedItem.getImageUrl());
-            photoRef.delete();
+            if (mDeletedItem.getImageUrl().contains("https://firebasestorage")) {
+                StorageReference photoRef = mFirebaseStorage.getReferenceFromUrl(
+                        mDeletedItem.getImageUrl());
+                photoRef.delete();
+            }
         }
     }
